refactor(multisig-prover): use non-zero nonce verifier set test data for abi encoding
