+++ contracts/multisig-prover/src/encoding/abi.rs
@@ -1,17 +1,78 @@
 use axelar_wasm_std::hash::Hash;
 use cosmwasm_std::HexBinary;
-use error_stack::ResultExt;
+use error_stack::{Result, ResultExt};
 use ethers_contract::contract::EthCall;
-use ethers_core::abi::{self, Tokenize};
-use evm_gateway::{ApproveMessagesCall, Message, Proof, RotateSignersCall, WeightedSigners};
+use ethers_core::abi::{encode as abi_encode, Token, Tokenize};
+use evm_gateway::{
+    ApproveMessagesCall, CommandType, Message, Proof, RotateSignersCall, WeightedSigners,
+};
+use itertools::Itertools;
 use multisig::msg::SignerWithSig;
 use multisig::verifier_set::VerifierSet;
+use sha3::{Digest, Keccak256};
 
 use crate::encoding::{to_recoverable, Encoder};
 use crate::error::ContractError;
 use crate::payload::Payload;
 
-pub fn encode(
+impl From<&Payload> for CommandType {
+    fn from(payload: &Payload) -> Self {
+        match payload {
+            Payload::Messages(_) => CommandType::ApproveMessages,
+            Payload::VerifierSet(_) => CommandType::RotateSigners,
+        }
+    }
+}
+
+pub fn payload_digest(
+    domain_separator: &Hash,
+    signer: &VerifierSet,
+    payload: &Payload,
+) -> Result<Hash, ContractError> {
+    let signer_hash = WeightedSigners::try_from(signer)
+        .map(|signers| signers.hash())
+        .change_context(ContractError::InvalidVerifierSet)?;
+
+    let data_hash = Keccak256::digest(encode_payload(payload)?);
+
+    // Prefix for standard EVM signed data https://eips.ethereum.org/EIPS/eip-191
+    let unsigned = [
+        "\x19Ethereum Signed Message:\n96".as_bytes(),
+        domain_separator,
+        signer_hash.as_slice(),
+        data_hash.as_slice(),
+    ]
+    .concat();
+
+    Ok(Keccak256::digest(unsigned).into())
+}
+
+pub fn encode_payload(payload: &Payload) -> Result<Vec<u8>, ContractError> {
+    let command_type = CommandType::from(payload).into();
+
+    match payload {
+        Payload::Messages(messages) => {
+            let messages = messages
+                .iter()
+                .map(Message::try_from)
+                .map_ok(|m| Token::Tuple(m.into_tokens()))
+                .collect::<Result<Vec<_>, _>>()
+                .change_context(ContractError::InvalidMessage)?;
+
+            Ok(abi_encode(&[command_type, Token::Array(messages)]))
+        }
+        Payload::VerifierSet(verifier_set) => Ok(abi_encode(&[
+            command_type,
+            Token::Tuple(
+                WeightedSigners::try_from(verifier_set)
+                    .change_context(ContractError::InvalidVerifierSet)?
+                    .into_tokens(),
+            ),
+        ])),
+    }
+}
+
+pub fn encode_execute_data(
     verifier_set: &VerifierSet,
     signers: Vec<SignerWithSig>,
     payload_digest: &Hash,
@@ -31,7 +92,7 @@ pub fn encode(
 
             (
                 ApproveMessagesCall::selector(),
-                abi::encode(&ApproveMessagesCall { messages, proof }.into_tokens()),
+                abi_encode(&ApproveMessagesCall { messages, proof }.into_tokens()),
             )
         }
         Payload::VerifierSet(new_verifier_set) => {
@@ -40,7 +101,7 @@ pub fn encode(
 
             (
                 RotateSignersCall::selector(),
-                abi::encode(&RotateSignersCall { new_signers, proof }.into_tokens()),
+                abi_encode(&RotateSignersCall { new_signers, proof }.into_tokens()),
             )
         }
     };
@@ -67,23 +128,64 @@ mod tests {
     use k256::ecdsa::Signature as K256Signature;
     use multisig::key::{KeyType, KeyTyped, Signature};
     use multisig::msg::{Signer, SignerWithSig};
-    use sha3::{Digest, Keccak256};
 
-    use crate::encoding::abi::execute_data::encode;
-    use crate::encoding::abi::payload_digest;
+    use crate::encoding::abi::{encode_execute_data, payload_digest, CommandType};
     use crate::encoding::add_27;
     use crate::payload::Payload;
     use crate::test::test_data::{
-        curr_verifier_set, domain_separator, messages, verifier_set_from_pub_keys,
+        curr_verifier_set, domain_separator, messages, new_verifier_set, verifier_set_from_pub_keys,
     };
 
     #[test]
-    fn rotate_signers_function_data() {
-        // RotateSigners function calldata hash generated by axelar-gmp-sdk-solidity unit tests
-        let expected_data_hash =
-            HexBinary::from_hex("52c65e01d464f6e440ebff7561b09edc1c9ff7754dd8bdaaa0410a952b9824cb")
-                .unwrap();
+    fn command_type_from_payload() {
+        let payload = Payload::Messages(vec![]);
+        assert_eq!(CommandType::from(&payload), CommandType::ApproveMessages);
 
+        let payload = Payload::VerifierSet(new_verifier_set());
+        assert_eq!(CommandType::from(&payload), CommandType::RotateSigners);
+    }
+
+    #[test]
+    fn abi_verifier_set_payload_digest() {
+        let domain_separator = domain_separator();
+
+        let new_pub_keys = vec![
+            "02de8b0cc10de1becab121cb1254a7b4075866b6e040d5a4cdd38c7ea6c68c7d0a",
+            "025a08780e7b80e64511006ec4db4128e18b31f05e9c8a4ef285322991d5f17332",
+            "03935a5be97cf2148cb5cb88d5f097a235859a572f46e53da907e80fd5578f9243",
+            "02515a95a89320988ff96f5e990b6d4c0a6807072f9b01c9ae634cf846bae2bd08",
+            "02464111b31e5d174ec44c172f5e3913d0a35344ef6c2cd8215494f23648ec3420",
+        ];
+
+        let mut new_verifier_set = verifier_set_from_pub_keys(new_pub_keys);
+        new_verifier_set.created_at = 2024;
+
+        goldie::assert!(hex::encode(
+            payload_digest(
+                &domain_separator,
+                &curr_verifier_set(),
+                &Payload::VerifierSet(new_verifier_set),
+            )
+            .unwrap()
+        ));
+    }
+
+    #[test]
+    fn abi_approve_messages_payload_digest() {
+        let domain_separator = domain_separator();
+
+        goldie::assert!(hex::encode(
+            payload_digest(
+                &domain_separator,
+                &curr_verifier_set(),
+                &Payload::Messages(messages()),
+            )
+            .unwrap()
+        ));
+    }
+
+    #[test]
+    fn abi_rotate_signers_execute_data() {
         let domain_separator = domain_separator();
 
         let new_pub_keys = vec![
@@ -94,7 +196,9 @@ mod tests {
             "03c55d66787c66f37257ef4b320ddcb64623d59e9bf0f3ec0f8ac7311b70cdd2c8",
         ];
 
-        let new_verifier_set = verifier_set_from_pub_keys(new_pub_keys);
+        let mut new_verifier_set = verifier_set_from_pub_keys(new_pub_keys);
+        new_verifier_set.created_at = 2024;
+
         let verifier_set = curr_verifier_set();
 
         // Generated signatures are already sorted by weight and evm address
@@ -113,11 +217,45 @@ mod tests {
             .try_into()
             .unwrap();
 
-        let execute_data =
-            encode(&verifier_set, signers_with_sigs, &payload_hash, &payload).unwrap();
-        let data_hash = Keccak256::digest(execute_data.as_slice());
+        goldie::assert!(encode_execute_data(
+            &verifier_set,
+            signers_with_sigs,
+            &payload_hash,
+            &payload
+        )
+        .unwrap()
+        .to_hex());
+    }
 
-        assert_eq!(HexBinary::from(data_hash.as_slice()), expected_data_hash);
+    #[test]
+    fn abi_approve_messages_execute_data() {
+        let domain_separator = domain_separator();
+        let verifier_set = curr_verifier_set();
+
+        // Generated signatures are already sorted by weight and evm address
+        let sigs: Vec<_> = vec![
+            "756473c3061df7ea3fef7c52e0e875dca2c93f08ce4f1d33e694d64c713a56842017d92f0a1b796afe1c5343677ff261a072fb210ff3d43cc2784c0774d4da7b1b",
+            "5bdad2b95e700283402392a2f5878d185f92d588a6b4868460977c4f06f4216f0452c2e215c2878fe6e146db5b74f278716a99b418c6b2cb1d812ad28e686cd81c",
+            "4c9c52a99a3941a384c4a80b3c5a14c059020d3d2f29be210717bdb9270ed55937fcec824313c90c198188ea8eb3b47c2bafe5e96c11f79ec793d589358024191b",
+        ].into_iter().map(|sig| HexBinary::from_hex(sig).unwrap()).collect();
+
+        let signers_with_sigs = signers_with_sigs(verifier_set.signers.values(), sigs);
+
+        let payload = Payload::Messages(messages());
+        let payload_hash: Hash = payload_digest(&domain_separator, &verifier_set, &payload)
+            .unwrap()
+            .as_slice()
+            .try_into()
+            .unwrap();
+
+        goldie::assert!(encode_execute_data(
+            &verifier_set,
+            signers_with_sigs,
+            &payload_hash,
+            &payload
+        )
+        .unwrap()
+        .to_hex());
     }
 
     #[test]
@@ -157,39 +295,6 @@ mod tests {
         } else {
             panic!("Invalid signature type")
         }
-    }
-
-    #[test]
-    fn approve_messages_function_data() {
-        // ApproveMessages function calldata hash generated by axelar-gmp-sdk-solidity unit tests
-        let expected_data_hash =
-            HexBinary::from_hex("973b422c8af25ef10a916130e2918667c15f791234a7ed449dbeae3cf9129f5d")
-                .unwrap();
-
-        let domain_separator = domain_separator();
-        let verifier_set = curr_verifier_set();
-
-        // Generated signatures are already sorted by weight and evm address
-        let sigs: Vec<_> = vec![
-            "756473c3061df7ea3fef7c52e0e875dca2c93f08ce4f1d33e694d64c713a56842017d92f0a1b796afe1c5343677ff261a072fb210ff3d43cc2784c0774d4da7b1b",
-            "5bdad2b95e700283402392a2f5878d185f92d588a6b4868460977c4f06f4216f0452c2e215c2878fe6e146db5b74f278716a99b418c6b2cb1d812ad28e686cd81c",
-            "4c9c52a99a3941a384c4a80b3c5a14c059020d3d2f29be210717bdb9270ed55937fcec824313c90c198188ea8eb3b47c2bafe5e96c11f79ec793d589358024191b",
-        ].into_iter().map(|sig| HexBinary::from_hex(sig).unwrap()).collect();
-
-        let signers_with_sigs = signers_with_sigs(verifier_set.signers.values(), sigs);
-
-        let payload = Payload::Messages(messages());
-        let payload_hash: Hash = payload_digest(&domain_separator, &verifier_set, &payload)
-            .unwrap()
-            .as_slice()
-            .try_into()
-            .unwrap();
-
-        let execute_data =
-            encode(&verifier_set, signers_with_sigs, &payload_hash, &payload).unwrap();
-        let data_hash = Keccak256::digest(execute_data.as_slice());
-
-        assert_eq!(HexBinary::from(data_hash.as_slice()), expected_data_hash);
     }
 
     fn signers_with_sigs<'a>(
