+++ contracts/multisig-prover/src/encoding/testdata/abi_approve_messages_payload_digest.golden
@@ -1,0 +1,1 @@
+1ce5dd08af564fc40645a826aeacb9d6daf005d81888fbadcf370682d46c6f22
