+++ contracts/multisig-prover/src/encoding/testdata/abi_rotate_signers_execute_data.golden
@@ -1,0 +1,1 @@
+1d92c0bf000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000000000000000000000000000000000007e80000000000000000000000000000000000000000000000000000000000000005000000000000000000000000126d4f2f15c471053968bdc5d53fe247499e2909000000000000000000000000000000000000000000000000000000000000000100000000000000000000000032ef31fbaea000336992ba75bd17ef632c3f55750000000000000000000000000000000000000000000000000000000000000001000000000000000000000000597d565dfe207dd6bc93ea291cbcc029b00c605a00000000000000000000000000000000000000000000000000000000000000010000000000000000000000007c53e31d6622e6dc761a029b5a1295afc98cda710000000000000000000000000000000000000000000000000000000000000001000000000000000000000000ef02cff50b393f6b233d4c50e88346af920825a4000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000500000000000000000000000015d34aaf54267db7d7c367839aaf71a00a2c6a6500000000000000000000000000000000000000000000000000000000000000010000000000000000000000003c44cdddb6a900fa2b585dd299e03d12fa4293bc000000000000000000000000000000000000000000000000000000000000000100000000000000000000000070997970c51812dc3a010c7d01b50e0d17dc79c8000000000000000000000000000000000000000000000000000000000000000100000000000000000000000090f79bf6eb2c4f870365e785982e1f101e93b9060000000000000000000000000000000000000000000000000000000000000001000000000000000000000000f39fd6e51aad88f6f4ce6ab8827279cfffb9226600000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000001600000000000000000000000000000000000000000000000000000000000000041e3a7c09bfa26df8bbd207df89d7ba01100b809324b2987e1426081284a50485345a5a20b6d1d5844470513099937f1015ce8f4832d3df97d053f044103434d8c1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041895dacfb63684da2360394d5127696129bd0da531d6877348ff840fb328297f870773df3c259d15dd28dbd51d87b910e4156ff2f3c1dc5f64d337dea7968a9401b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000417c685ecc8a42da4cd9d6de7860b0fddebb4e2e934357500257c1070b1a15be5e27f13b627cf9fa44f59d535af96be0a5ec214d988c48e2b5aaf3ba537d0215bb1b00000000000000000000000000000000000000000000000000000000000000
