+++ contracts/multisig-prover/src/encoding/testdata/abi_verifier_set_payload_digest.golden
@@ -1,0 +1,1 @@
+79be4b94bc9da6fef155bfc24710cfadae6935f7de332e2374c6176fba683de0
