+++ contracts/multisig-prover/src/testdata/payload_verifier_set_id_unchanged.golden
@@ -1,1 +1,1 @@
-"76dc1094e9f8dbb6a98c077e3a67c889d49820e98e8560f0a91e3531d41d4168"
+"52a44500092c6ff48468b77dbc926bfd05e6f788a3ad79a6ea68b52beeca93ce"
